fix(phonebook): use administrator user
The pbook2webtop script was assuming the main
user of the domain was named 'admin' as for NS7.
But the new default for NS8 is 'administrator'.

Add a logic to discover the current real administator
user. If no user is found, default is 'administrator'.

Also make the script configurabile from env variable:
- PHONEBOOK_WEBTOP_ADMIN: override the discovered user
- PHONEBOOK_WBTOP_FOLDER: override contact category name,
  default to 'Rubrica centralizzata'